--- src/main/java/com/jumpcutfindo/happyholidays/common/block/christmas/gifts/GiftWrapperBlock.java
@@ -1,121 +1,0 @@
-package com.jumpcutfindo.happyholidays.common.block.christmas.gifts;
-
-import javax.annotation.Nullable;
-
-import com.jumpcutfindo.happyholidays.HappyHolidaysMod;
-import com.jumpcutfindo.happyholidays.common.block.christmas.ChristmasBlock;
-import com.jumpcutfindo.happyholidays.common.registry.TileEntityRegistry;
-import com.jumpcutfindo.happyholidays.common.tileentity.christmas.GiftWrapperTileEntity;
-
-import net.minecraft.block.AbstractBlock;
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockState;
-import net.minecraft.block.SoundType;
-import net.minecraft.block.material.Material;
-import net.minecraft.client.renderer.RenderType;
-import net.minecraft.client.renderer.RenderTypeLookup;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.entity.player.ServerPlayerEntity;
-import net.minecraft.inventory.InventoryHelper;
-import net.minecraft.item.BlockItemUseContext;
-import net.minecraft.item.Item;
-import net.minecraft.state.DirectionProperty;
-import net.minecraft.state.StateContainer;
-import net.minecraft.state.properties.BlockStateProperties;
-import net.minecraft.tileentity.TileEntity;
-import net.minecraft.util.ActionResultType;
-import net.minecraft.util.Direction;
-import net.minecraft.util.Hand;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.BlockRayTraceResult;
-import net.minecraft.util.math.shapes.ISelectionContext;
-import net.minecraft.util.math.shapes.VoxelShape;
-import net.minecraft.util.math.shapes.VoxelShapes;
-import net.minecraft.world.IBlockReader;
-import net.minecraft.world.World;
-import net.minecraftforge.fml.network.NetworkHooks;
-
-public class    GiftWrapperBlock extends ChristmasBlock {
-    public static final DirectionProperty FACING = BlockStateProperties.HORIZONTAL_FACING;
-
-    public static final String BLOCK_ID = "gift_wrapping_station";
-
-    public static final Properties BLOCK_PROPERTIES =
-            AbstractBlock.Properties
-                    .of(Material.WOOD)
-                    .harvestLevel(-1)
-                    .strength(0.5f)
-                    .sound(SoundType.WOOD)
-                    .noOcclusion();
-
-    public static final Item.Properties ITEM_PROPERTIES =
-            new Item.Properties().tab(HappyHolidaysMod.HAPPY_HOLIDAYS_GROUP);
-
-    public static final VoxelShape SHAPE = VoxelShapes.or(
-            Block.box(2.0, 0.0, 2.0, 14.0, 2.0, 14.0),
-            Block.box(5.0, 2.0, 5.0, 11.0, 12.0, 11.0),
-            Block.box(0.0, 12.0, 0.0, 16.0, 16.0, 16.0)
-    );
-
-    public GiftWrapperBlock() {
-        super(BLOCK_PROPERTIES);
-        this.registerDefaultState(this.getStateDefinition().any()
-                .setValue(FACING, Direction.NORTH)
-        );
-    }
-
-    protected void createBlockStateDefinition(StateContainer.Builder<Block, BlockState> stateBuilder) {
-        stateBuilder.add(FACING);
-    }
-
-    @Override
-    public void configureBlock() {
-        RenderTypeLookup.setRenderLayer(this, RenderType.cutoutMipped());
-    }
-
-    @Override
-    public boolean hasTileEntity(BlockState state) {
-        return true;
-    }
-
-    @Nullable
-    @Override
-    public TileEntity createTileEntity(BlockState state, IBlockReader world) {
-        return TileEntityRegistry.GIFT_WRAPPER_ENTITY_TYPE.get().create();
-    }
-
-    @Override
-    public ActionResultType use(BlockState blockState, World world, BlockPos blockPos,
-                                PlayerEntity playerEntity, Hand hand, BlockRayTraceResult rayTraceResult) {
-        if (world.isClientSide()) return ActionResultType.SUCCESS;
-        else {
-            TileEntity te = world.getBlockEntity(blockPos);
-            if (te instanceof GiftWrapperTileEntity) {
-                NetworkHooks.openGui((ServerPlayerEntity) playerEntity, (GiftWrapperTileEntity) te, blockPos);
-            }
-            return ActionResultType.CONSUME;
-        }
-    }
-
-    @Override
-    public void playerWillDestroy(World world, BlockPos blockPos, BlockState blockState,
-                                  PlayerEntity playerEntity) {
-        super.playerWillDestroy(world, blockPos, blockState, playerEntity);
-        TileEntity te = world.getBlockEntity(blockPos);
-        if (te instanceof GiftWrapperTileEntity) {
-            InventoryHelper.dropContents(world, blockPos, (GiftWrapperTileEntity) te);
-        }
-    }
-
-
-    @Override
-    public VoxelShape getShape(BlockState blockState, IBlockReader blockReader, BlockPos blockPos,
-                               ISelectionContext context) {
-        return SHAPE;
-    }
-
-    @Nullable
-    public BlockState getStateForPlacement(BlockItemUseContext context) {
-        return this.defaultBlockState().setValue(FACING, context.getHorizontalDirection().getOpposite());
-    }
-}
